requirements.txt
# =============================================================================
# PSVPP (Platform Supply Vessel Planning Problem) - Requirements
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies - Essential packages for basic functionality
# -----------------------------------------------------------------------------
numpy>=1.21.0                    # Numerical computing
pandas>=1.3.0                    # Data manipulation and analysis
pyyaml>=6.0                      # YAML configuration parsing

# -----------------------------------------------------------------------------
# Visualization & UI - Plotting and dashboard components
# -----------------------------------------------------------------------------
matplotlib>=3.5.0                # Static plotting
seaborn>=0.11.0                  # Statistical data visualization
plotly>=5.0.0                    # Interactive plotting
dash>=2.0.0                      # Web-based dashboards

# -----------------------------------------------------------------------------
# Reinforcement Learning - RL environment and training
# -----------------------------------------------------------------------------
gymnasium>=0.28.0                # RL environment standard
stable-baselines3[extra]>=2.0.0  # RL algorithms (includes extras)
torch>=1.11.0                    # Deep learning framework

# -----------------------------------------------------------------------------
# Rust Integration - Python-Rust bindings
# -----------------------------------------------------------------------------
maturin>=1.0.0                   # Rust-Python extension builder

# -----------------------------------------------------------------------------
# Development & Testing - Code quality and testing tools
# -----------------------------------------------------------------------------
pytest>=7.0.0                    # Testing framework
black>=22.0.0                    # Code formatter
mypy>=0.950                      # Static type checking

# -----------------------------------------------------------------------------
# Optional: Advanced Features (uncomment as needed)
# -----------------------------------------------------------------------------
# ray[rllib]>=2.0.0              # Distributed RL training
# tensorboard>=2.8.0             # Training visualization
# wandb>=0.12.0                  # Experiment tracking
# jupyter>=1.0.0                 # Notebook environment
# networkx>=2.6.0                # Graph algorithms
# scikit-learn>=1.0.0            # Machine learning utilities
